requirements.txt
streamlit
pandas
google-auth-oauthlib
google-cloud-firestore